+++ apps/angular-example/src/app/abby.ts
@@ -1,5 +1,12 @@
 import { Injectable } from "@angular/core";
-import { AbbyService, InferFlagNames, InferFlags, InferTestNames, InferTests } from "abby";
+import {
+  AbbyService,
+  InferFlagNames,
+  InferRemoteConfig,
+  InferRemoteConfigName,
+  InferTestNames,
+  InferTests,
+} from "abby";
 import abby from "../../abby.config";
 
 @Injectable({
@@ -10,5 +17,6 @@ export class Abby extends AbbyService<
   InferFlagNames<typeof abby>,
   InferTestNames<typeof abby>,
   InferTests<typeof abby>,
-  InferFlags<typeof abby>
+  InferRemoteConfig<typeof abby>,
+  InferRemoteConfigName<typeof abby>
 > {}