+++ packages/angular/src/lib/StorageService.ts
@@ -1,6 +1,7 @@
 import {
   getABStorageKey,
   getFFStorageKey,
+  getRCStorageKey,
   type IStorageService,
 } from "@tryabby/core";
 
@@ -62,5 +63,34 @@ class FFStorageService implements IStorageService {
   }
 }
 
+class RCStorageService implements IStorageService {
+  get(projectId: string, key: string): string | null {
+    const name = `${getRCStorageKey(projectId, key)}=`;
+    const decodedCookie = decodeURIComponent(document.cookie);
+    const cookieArray = decodedCookie.split(";");
+
+    for (let i = 0; i < cookieArray.length; i++) {
+      let cookie = cookieArray[i];
+      while (cookie.charAt(0) === " ") {
+        cookie = cookie.substring(1);
+      }
+      if (cookie.indexOf(name) === 0) {
+        return cookie.substring(name.length, cookie.length);
+      }
+    }
+    return null;
+  }
+  set(projectId: string, key: string, value: string): void {
+    const token = getRCStorageKey(projectId, key);
+    document.cookie = `${token}=${value}`;
+  }
+
+  remove(projectId: string, key: string): void {
+    const token = getRCStorageKey(projectId, key);
+    document.cookie = `${token}=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;`;
+  }
+}
+
 export const TestStorageService = new ABStorageService();
 export const FlagStorageService = new FFStorageService();
+export const RemoteConfigStorageService = new RCStorageService();