+++ packages/core/src/shared/constants.ts
@@ -2,4 +2,6 @@ export const DOCS_URL = "https://docs.tryabby.com/";
 export const ABBY_BASE_URL = "https://www.tryabby.com/";
 export const ABBY_AB_STORAGE_PREFIX = "__abby__ab__";
 export const ABBY_FF_STORAGE_PREFIX = "__abby__ff__";
+export const ABBY_RC_STORAGE_PREFIX = "__abby__rc__";
 export const ABBY_INSTANCE_KEY = "__abby_instance__";
+export const DEFAULT_FEATURE_FLAG_VALUE = false;