+++ packages/core/src/shared/helpers.ts
@@ -1,5 +1,9 @@
-import { ABBY_AB_STORAGE_PREFIX, ABBY_FF_STORAGE_PREFIX } from "./constants";
-import { FlagValue, FlagValueString } from "./schemas";
+import {
+  ABBY_AB_STORAGE_PREFIX,
+  ABBY_FF_STORAGE_PREFIX,
+  ABBY_RC_STORAGE_PREFIX,
+} from "./constants";
+import { RemoteConfigValue, RemoteConfigValueString } from "./schemas";
 
 export function getABStorageKey(projectId: string, testName: string): string {
   return `${ABBY_AB_STORAGE_PREFIX}${projectId}_${testName}`;
@@ -9,20 +13,22 @@ export function getFFStorageKey(projectId: string, flagName: string): string {
   return `${ABBY_FF_STORAGE_PREFIX}${projectId}_${flagName}`;
 }
 
+export function getRCStorageKey(projectId: string, remoteConfigName: string): string {
+  return `${ABBY_RC_STORAGE_PREFIX}${projectId}_${remoteConfigName}`;
+}
+
 export function assertUnreachable(x: never): never {
   throw new Error("Reached unreachable code");
 }
 
-export function flagStringToType({
+export function remoteConfigStringToType({
   stringifiedValue,
-  flagType,
+  remoteConfigType,
 }: {
   stringifiedValue: string;
-  flagType: FlagValueString;
-}): FlagValue {
-  switch (flagType) {
-    case "Boolean":
-      return stringifiedValue === "true";
+  remoteConfigType: RemoteConfigValueString;
+}): RemoteConfigValue {
+  switch (remoteConfigType) {
     case "String":
       return stringifiedValue;
     case "Number":
@@ -30,14 +36,14 @@ export function flagStringToType({
     case "JSON":
       return JSON.parse(stringifiedValue);
     default:
-      assertUnreachable(flagType);
+      assertUnreachable(remoteConfigType);
   }
 }
 
-export function getDefaultFlagValue(flagType: FlagValueString): FlagValue {
-  switch (flagType) {
-    case "Boolean":
-      return false;
+export function getDefaultRemoteConfigValue(
+  remoteConfigType: RemoteConfigValueString
+): RemoteConfigValue {
+  switch (remoteConfigType) {
     case "String":
       return "";
     case "Number":
@@ -45,13 +51,12 @@ export function getDefaultFlagValue(flagType: FlagValueString): FlagValue {
     case "JSON":
       return {};
     default:
-      assertUnreachable(flagType);
+      assertUnreachable(remoteConfigType);
   }
 }
 
-export function stringifyFlagValue(value: FlagValue) {
+export function stringifyRemoteConfigValue(value: RemoteConfigValue) {
   switch (typeof value) {
-    case "boolean":
     case "number":
       return value.toString();
     case "string":