+++ packages/core/tests/types.test.ts
@@ -1,15 +1,12 @@
-import { AbbyConfig, FlagValueStringToType } from "../dist";
-import { Abby } from "../src/index";
+import { RemoteConfigValueStringToType } from "../dist";
+import { Abby, AbbyConfig } from "../src/index";
 
 describe("types", () => {
   it("produces proper types", () => {
     const abby = new Abby({
       environments: [""],
       projectId: "",
-      flags: {
-        flag1: "Boolean",
-        flag2: "String",
-      },
+      flags: ["flag1"],
       tests: {
         test1: {
           variants: ["firstVariant", "secondVariant"],
@@ -18,11 +15,13 @@ describe("types", () => {
           variants: ["some-random-string", "abc"],
         },
       },
+      remoteConfig: { remoteConfig1: "String" },
     });
 
     assertType<boolean>(abby.getFeatureFlag("flag1"));
-    assertType<string>(abby.getFeatureFlag("flag2"));
-    expectTypeOf(abby.getFeatureFlag).parameter(0).toEqualTypeOf<"flag1" | "flag2">();
+    assertType<string>(abby.getRemoteConfig("remoteConfig1"));
+    expectTypeOf(abby.getFeatureFlag).parameter(0).toEqualTypeOf<"flag1">();
+    expectTypeOf(abby.getRemoteConfig).parameter(0).toEqualTypeOf<"remoteConfig1">();
 
     assertType<"firstVariant" | "secondVariant">(abby.getTestVariant("test1"));
 
@@ -31,30 +30,20 @@ describe("types", () => {
     // This is the potential type of the devOverrides
     type DevOverrides = NonNullable<
       NonNullable<
-        NonNullable<
-          AbbyConfig<
-            "flag1" | "flag2",
-            {},
-            {
-              flag1: "Boolean";
-              flag2: "String";
-            }
-          >["settings"]
-        >["flags"]
+        NonNullable<AbbyConfig<"flag1", {}, ["flag1"]>["settings"]>["flags"]
       >["devOverrides"]
     >;
 
     expectTypeOf<DevOverrides>().toEqualTypeOf<{
       flag1?: boolean;
-      flag2?: string;
     }>();
   });
 });
 
 describe("Type Helpers", () => {
   it("converts the strings properly", () => {
-    expectTypeOf<FlagValueStringToType<"String">>().toEqualTypeOf<string>();
-    expectTypeOf<FlagValueStringToType<"Boolean">>().toEqualTypeOf<boolean>();
-    expectTypeOf<FlagValueStringToType<"Number">>().toEqualTypeOf<number>();
+    expectTypeOf<RemoteConfigValueStringToType<"String">>().toEqualTypeOf<string>();
+    expectTypeOf<RemoteConfigValueStringToType<"JSON">>().toEqualTypeOf<Record<string, unknown>>();
+    expectTypeOf<RemoteConfigValueStringToType<"Number">>().toEqualTypeOf<number>();
   });
 });