+++ apps/angular-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # angular-example
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/devtools@5.0.0
+  - @tryabby/angular@2.0.0
+
 ## 0.0.9
 
 ### Patch Changes