+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @tryabby/angular
 
+## 2.0.0
+
+### Major Changes
+
+- add remote config
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/core@5.0.0
+
 ## 1.2.0
 
 ### Minor Changes